requirements.txt
aiohttp>=3.8.1
pillow>=10.3.0